Add files via upload
This script contains 2 options, landing, and airtime.

Landing will search for the earliest landing.

Airtime will search for total time spent in the air, could be buggy.

This script also contains a position condition to check if the car is too much to the left or right. Useful for when you need to avoid adding wall barriers in the current map. (This if condition is under def on_bruteforce_evaluate function)